requirements.txt
jira
PyGithub
python-dotenv
requests
pytest
